# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pet Spa Booking API Documentation",
    "description": "RESTful API cho hệ thống đặt lịch Pet Spa",
    "version": "1.0.0",
    "contact": {
      "name": "Pet Spa Support",
      "email": "admin@petspa.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Pet Spa Booking API Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer token"
      }
    },
    "schemas": {
...
      "Service": {
        "type": "object",
        "properties": {
          "serviceID": {"type": "integer", "example": 1},
          "serviceName": {"type": "string", "example": "Tắm và sấy khô"},
          "description": {"type": "string", "example": "Dịch vụ tắm spa cao cấp"},
          "price": {"type": "number", "example": 150000},
          "duration": {"type": "integer", "example": 60},
          "categoryID": {"type": "integer", "example": 1},
          "image": {"type": "string", "example": "services/image.jpg"}
        }
      },
...
    }
  },
  "paths": {
...
    "/services": {
      "get": {
        "tags": ["Services"],
        "summary": "Lấy danh sách dịch vụ",
        "parameters": [
          {"name": "search", "in": "query", "description": "Tìm kiếm theo tên hoặc mô tả", "schema": {"type": "string"}},
          {"name": "category_id", "in": "query", "description": "Lọc theo danh mục", "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean", "example": true},
                    "data": {"type": "array", "items": {"$ref": "#/components/schemas/Service"}}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "tags": ["Services"],
        "summary": "Lấy chi tiết dịch vụ",
        "parameters": [
          {"name": "id", "in": "path", "required": true, "description": "ID dịch vụ", "schema": {"type": "integer"}}
        ],
        "responses": {
          "200": {"description": "Thành công"},
          "404": {"description": "Không tìm thấy"}
        }
      }
    },
...
    "/services/calculate-price": {
      "post": {
        "tags": ["Services"],
        "summary": "Tính giá dịch vụ theo thú cưng",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["service_id", "pet_id"],
                "properties": {
                  "service_id": {"type": "integer", "example": 1},
                  "pet_id": {"type": "integer", "example": 1}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Tính giá thành công"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/profile": {
      "get": {
        "tags": ["Profile"],
        "summary": "Lấy thông tin profile",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "Thành công"},
          "401": {"description": "Unauthorized"}
        }
      },
      "put": {
        "tags": ["Profile"],
        "summary": "Cập nhật thông tin profile",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email"],
                "properties": {
                  "name": {"type": "string", "example": "Nguyễn Văn A"},
                  "email": {"type": "string", "format": "email"},
                  "phone": {"type": "string"},
                  "address": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Cập nhật thành công"},
          "401": {"description": "Unauthorized"}
        }
      },
      "delete": {
        "tags": ["Profile"],
        "summary": "Xóa tài khoản",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": {"type": "string", "format": "password"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Xóa thành công"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
    "/profile/password": {
      "put": {
        "tags": ["Profile"],
        "summary": "Đổi mật khẩu",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["current_password", "password", "password_confirmation"],
                "properties": {
                  "current_password": {"type": "string", "format": "password"},
                  "password": {"type": "string", "format": "password"},
                  "password_confirmation": {"type": "string", "format": "password"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Đổi mật khẩu thành công"},
          "401": {"description": "Unauthorized"}
        }
      }
    },
...
  }
}